requirement.txt
django
djangorestframework
djangorestframework-simplejwt
mysqlclient
django-cors-headers
django-filter